programs.neovim: add options for RC header/footer and mapleader (#54)
When configuring plugins, one can't make use of `<leader>` because it's not possible to configure the `mapleader` ahead of the plugins. I've added options for adding random lines to the top/bottom of the rc file, and I added another option for just the mapleader character, and it itself will inject the code in the header.